# [aggregator] Avoid large copies in entry rollup comparisons