@robskillington 2021-02-07
Will this cause any new heap allocations due to escape analysis not able to see if ptr is held onto from a function called down the stack? `go build -gcflags="-m"` should be able to show or not I believe.