@vdarulis 2021-02-09
> Will this cause any new heap allocations due to escape analysis not able to see if ptr is held onto from a function called down the stack? `go build -gcflags="-m"` should be able to show or not I believe.

@robskillington it did not, but in the end went for a different approach, as the root cause of all these copies is Go inlining, or rather lack of it - dropping this in favor of #3195, which does not change semantics and is actually even faster.